Remove threaded port updates
They are a big headache for not much benefit. rock-display itself
has no use for it(every change touches UI, so must happen on
that thread), plugins can create their own readers and poll them
as they like.

As for the headache: When passing pointers into the samples through
the qt signal system, one must cope with the caller regaining control
before the callee sees the data, or always use DirectConnection or
BlockingQueuedConnection. DirectConnection conflicts with UI updates
only possible on UI thread, BlockingQueuedConnection basically
synchronizes the updating thread to the UI thread. This could be
worked around by always creating a new sample and destroying it
once all users are done, but then, one would be creating
lots of structures, and run in a synchronisation problem when multiple
writers exist in rock-display(like the configitemhandlers or editing
plugins), possibly working on stale copies.